requirements/base.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
python-dotenv
alembic
